fix win32 warning about subtraction by InterlockedExchangeAdd
It complains about overflow. It probably means that the result can be
unsigned.

+++ util.h
@@ -17,7 +17,7 @@
 
 #define sleep(seconds) Sleep((seconds) * 1000)
 #define __sync_fetch_and_add(ap, b) InterlockedExchangeAdd((LONG volatile *) (ap), (b));
-#define __sync_fetch_and_sub(ap, b) InterlockedExchangeAdd((LONG volatile *) (ap), -(b));
+#define __sync_fetch_and_sub(ap, b) InterlockedExchangeAdd((LONG volatile *) (ap), -((LONG)b));
 
 /* MinGW gcc, no function prototype for built-in function stpcpy() */ 
 char *stpcpy (char *dest, const char *src);
